+++ packages/components/organisms/f-cookie-banner/test/component/f-cookieBanner-legacy.component.shared.spec.js
@@ -7,7 +7,7 @@ const cookieBanner = new CookieBanner('organism', 'cookie-banner-component');
 describe('Legacy - f-cookieBanner component tests - @browserstack', () => {
     it('should display the f-cookieBanner component', () => {
         // Arrange
-        cookieBanner.withQuery('&knob-Locale', 'en-GB');
+        cookieBanner.withQuery('&knob-Locale', 'en-AU');
         const pageUrl = buildUrl(cookieBanner.componentType, cookieBanner.componentName, cookieBanner.path);
 
         cookieBanner.open(pageUrl);