+++ packages/components/organisms/f-checkout/CHANGELOG.md
@@ -4,6 +4,34 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v0.143.1
+------------------------------
+*June 21, 2021*
+
+### Changed
+- Flipped coordinates from local storage
+
+
+v0.143.0
+------------------------------
+*June 18, 2021*
+
+### Changed
+- Fixed redirect link for Dine In orders
+
+
+v0.142.0
+------------------------------
+*June 17, 2021*
+
+### Changed
+- Percy to run desktop / mobile in separate specs
+
+
+### Fixed
+- Percy visual regression test
+
+
 v0.141.1
 ------------------------------
 *June 17, 2021*